+++ libs/features/mas/src/merch-card-collection.js
@@ -1,15 +1,21 @@
-import { html, LitElement } from 'lit';
+import { html, LitElement, css, unsafeCSS, nothing } from 'lit';
+import { DESKTOP_UP, TABLET_UP } from './media.js';
+import { MatchMediaController } from '@spectrum-web-components/reactive-controllers/src/MatchMedia.js';
+import { getCollectionOptions } from './variants/variants.js';
 import { deeplink, pushState } from './deeplink.js';
-import { EVENT_MAS_ERROR, EVENT_MERCH_CARD_COLLECTION_LITERALS_CHANGED, EVENT_MERCH_CARD_COLLECTION_SIDENAV_ATTACHED, EVENT_MERCH_SIDENAV_SELECT, SORT_ORDER } from './constants.js';
-
 import {
+    EVENT_MAS_ERROR,
+    EVENT_MERCH_CARD_COLLECTION_LITERALS_CHANGED,
+    EVENT_MERCH_CARD_COLLECTION_SIDENAV_ATTACHED,
+    EVENT_MERCH_SIDENAV_SELECT,
     EVENT_MERCH_CARD_COLLECTION_SORT,
     EVENT_MERCH_CARD_COLLECTION_SHOWMORE,
     EVENT_AEM_ERROR,
     EVENT_AEM_LOAD,
+    SORT_ORDER
 } from './constants.js';
 import { styles } from './merch-card-collection.css.js';
-import { getService } from './utils.js';
+import { getService, getSlotText } from './utils.js';
 import './mas-commerce-service';
 
 const MERCH_CARD_COLLECTION = 'merch-card-collection';
@@ -234,9 +240,7 @@ export class MerchCardCollection extends LitElement {
         this.stopDeeplink?.();
     }
 
-    async initializeHeader() {
-        await import('./merch-card-collection-header.js');
-        await customElements.whenDefined('merch-card-collection-header');
+    initializeHeader() {
         const header = document.createElement('merch-card-collection-header');
         header.collection = this;
         header.classList.add(this.variant);
@@ -265,15 +269,15 @@ export class MerchCardCollection extends LitElement {
         }
     }
 
-    async attachSidenav(sidenav, append = true) {
+    attachSidenav(sidenav, append = true) {
         if (!sidenav) return;
-        await this.initializeHeader();
         if (append) this.parentElement.prepend(sidenav);
         this.sidenav = sidenav;
         this.sidenav.variant = this.variant;
         this.sidenav.classList.add(this.variant);
         if (SIDENAV_AUTOCLOSE[this.variant]) 
             this.sidenav.setAttribute('autoclose', '');
+        this.initializeHeader();
         this.dispatchEvent(new CustomEvent(EVENT_MERCH_CARD_COLLECTION_SIDENAV_ATTACHED));
     }
 
@@ -474,3 +478,370 @@ export class MerchCardCollection extends LitElement {
 MerchCardCollection.SortOrder = SORT_ORDER;
 
 customElements.define(MERCH_CARD_COLLECTION, MerchCardCollection);
+
+// #region Header
+
+const RESULT_TEXT_SLOT_NAMES = {
+    // no search
+    filters: ['noResultText', 'resultText', 'resultsText'],
+    filtersMobile: ['noResultText', 'resultMobileText', 'resultsMobileText'],
+    // search on desktop
+    search: ['noSearchResultsText', 'searchResultText', 'searchResultsText'],
+    // search on mobile
+    searchMobile: [
+        'noSearchResultsMobileText',
+        'searchResultMobileText',
+        'searchResultsMobileText',
+    ],
+  };
+  
+  const updatePlaceholders = (el, key, value) => {
+    const placeholders = el.querySelectorAll(`[data-placeholder="${key}"]`);
+    placeholders.forEach(placeholder => {
+        placeholder.innerText = value || '';
+    });
+  };
+  
+  const defaultVisibility = {
+    search: ['mobile', 'tablet'],
+    filter: ['mobile', 'tablet'],
+    sort: true,
+    result: true,
+    custom: false,
+  }
+  
+  const SEARCH_SIZE = {
+      catalog: 'l'
+  };
+  
+  export default class MerchCardCollectionHeader extends LitElement {
+      constructor() {
+          super();
+          this.collection = null;
+          this.#visibility = {
+              search: false,
+              filter: false,
+              sort: false,
+              result: false,
+              custom: false
+          }
+          this.updateLiterals = this.updateLiterals.bind(this);
+          this.handleSidenavAttached = this.handleSidenavAttached.bind(this);
+      }
+  
+      #visibility;
+  
+      connectedCallback() {
+          super.connectedCallback();
+          this.collection?.addEventListener(EVENT_MERCH_CARD_COLLECTION_LITERALS_CHANGED, this.updateLiterals);
+          this.collection?.addEventListener(EVENT_MERCH_CARD_COLLECTION_SIDENAV_ATTACHED, this.handleSidenavAttached);
+      }
+  
+      disconnectedCallback() {
+          super.disconnectedCallback();
+          this.collection?.removeEventListener(EVENT_MERCH_CARD_COLLECTION_LITERALS_CHANGED, this.updateLiterals);
+          this.collection?.removeEventListener(EVENT_MERCH_CARD_COLLECTION_SIDENAV_ATTACHED, this.handleSidenavAttached);
+      }
+  
+      willUpdate() {
+          this.#visibility.search = this.getVisibility('search');
+          this.#visibility.filter = this.getVisibility('filter');
+          this.#visibility.sort = this.getVisibility('sort');
+          this.#visibility.result = this.getVisibility('result');
+          this.#visibility.custom = this.getVisibility('custom');
+      }
+  
+      parseVisibilityOptions(visibility, type) {
+          if (!visibility) return null;
+          if (!Object.hasOwn(visibility, type)) return null;
+          const typeVisibility = visibility[type];
+          if (typeVisibility === false) return false;
+          if (typeVisibility === true) return true;
+          return typeVisibility.includes(this.currentMedia);
+      }
+  
+      getVisibility(type) {
+          const visibility = getCollectionOptions(this.collection?.variant)?.headerVisibility;
+          const typeVisibility = this.parseVisibilityOptions(visibility, type);
+          if (typeVisibility !== null) return typeVisibility;
+          return this.parseVisibilityOptions(defaultVisibility, type);
+      }
+  
+      get sidenav() {
+          return this.collection?.sidenav;
+      }
+  
+      get search() {
+          return this.collection?.search;
+      }
+  
+      get resultCount() {
+          return this.collection?.resultCount;
+      }
+  
+      get variant() {
+          return this.collection?.variant;
+      }
+  
+      tablet = new MatchMediaController(this, TABLET_UP);
+      desktop = new MatchMediaController(this, DESKTOP_UP);
+  
+      get isMobile() {
+          return !this.isTablet && !this.isDesktop;
+      }
+  
+      get isTablet() {
+          return this.tablet.matches && !this.desktop.matches;
+      }
+  
+      get isDesktop() {
+          return this.desktop.matches;
+      }
+  
+      get currentMedia() {
+          if (this.isDesktop) return 'desktop';
+          if (this.isTablet) return 'tablet';
+          return 'mobile';
+      }
+  
+      get searchAction() {
+          if (!this.#visibility.search) return nothing;
+          const searchPlaceholder = getSlotText(this, 'searchText');
+          if (!searchPlaceholder) return nothing;
+          return html`
+              <merch-search deeplink="search" id="search">
+                  <sp-search
+                      id="search-bar"
+                      placeholder="${searchPlaceholder}"
+                      .size=${SEARCH_SIZE[this.variant]}
+                  ></sp-search>
+              </merch-search>
+          `;
+      }
+  
+      get filterAction() {
+          if (!this.#visibility.filter) return nothing;
+          if (!this.sidenav) return nothing;
+          return html`
+              <sp-action-button
+                id="filter"
+                variant="secondary"
+                treatment="outline"
+                @click="${this.openFilters}"
+                ><slot name="filtersText"></slot
+              ></sp-action-button>
+          `;
+      }
+  
+      get sortAction() {
+          if (!this.#visibility.sort) return nothing;
+          const sortText = getSlotText(this, 'sortText');
+          if (!sortText) return;
+          const popularityText = getSlotText(this, 'popularityText');
+          const alphabeticallyText = getSlotText(this, 'alphabeticallyText');
+  
+          if (!(popularityText && alphabeticallyText)) return;
+          const alphabetical = this.collection?.sort === SORT_ORDER.alphabetical;
+  
+          return html`
+              <sp-action-menu
+                  id="sort"
+                  size="m"
+                  @change="${this.collection?.sortChanged}"
+                  selects="single"
+                  value="${alphabetical
+                      ? SORT_ORDER.alphabetical
+                      : SORT_ORDER.authored}"
+              >
+                  <span slot="label-only"
+                      >${sortText}:
+                      ${alphabetical ? alphabeticallyText : popularityText}</span
+                  >
+                  <sp-menu-item value="${SORT_ORDER.authored}"
+                      >${popularityText}</sp-menu-item
+                  >
+                  <sp-menu-item value="${SORT_ORDER.alphabetical}"
+                      >${alphabeticallyText}</sp-menu-item
+                  >
+              </sp-action-menu>
+          `;
+      }
+  
+      get resultSlotName() {
+          const slotType = `${this.search ? 'search' : 'filters'}${this.isMobile || this.isTablet ? 'Mobile' : ''}`;
+          return RESULT_TEXT_SLOT_NAMES[slotType][Math.min(this.resultCount, 2)];
+      }
+  
+      get resultLabel() {
+          if (!this.#visibility.result) return nothing;
+          if (!this.sidenav) return nothing;
+          const type = this.search ? 'search' : 'filter';
+          const quantity = !this.resultCount ? 'none' : this.resultCount === 1 ? 'single' : 'multiple';
+          return html`
+            <div id="result" aria-live="polite" type=${type} quantity=${quantity}>
+                <slot name="${this.resultSlotName}"></slot>
+            </div>`
+      }
+  
+      get customArea() {
+          if (!this.#visibility.custom) return nothing;
+          const customHeaderAreaGetter = getCollectionOptions(this.collection?.variant)?.customHeaderArea;
+          if (!customHeaderAreaGetter) return nothing;
+          const customHeaderArea = customHeaderAreaGetter(this.collection);
+          if (!customHeaderArea || customHeaderArea === nothing) return nothing;
+          return html`<div id="custom">${customHeaderArea}</div>`;
+      }
+  
+      // #region Handlers
+  
+      openFilters(event) {
+          this.sidenav.showModal(event);
+      }
+  
+      updateLiterals(event) {
+          Object.keys(event.detail).forEach(key => {
+              updatePlaceholders(this, key, event.detail[key]);
+          });
+          this.requestUpdate();
+      }
+  
+      handleSidenavAttached() {
+          this.requestUpdate();
+      }
+  
+      // #endregion
+  
+      render() {
+          return html`
+            <sp-theme color="light" scale="medium">
+              <div id="header">${this.searchAction}${this.filterAction}${this.sortAction}${this.resultLabel}${this.customArea}</div>
+            </sp-theme>
+          `
+      }
+  
+      static styles = css`
+          :host {
+              --merch-card-collection-header-max-width: var(--merch-card-collection-card-width);
+              --merch-card-collection-header-margin-bottom: 32px;
+              --merch-card-collection-header-column-gap: 8px;
+              --merch-card-collection-header-row-gap: 16px;
+              --merch-card-collection-header-columns: auto auto;
+              --merch-card-collection-header-areas: "search search" 
+                                                    "filter sort"
+                                                    "result result";
+              --merch-card-collection-header-search-max-width: unset;
+              --merch-card-collection-header-filter-height: 40px;
+              --merch-card-collection-header-filter-font-size: 16px;
+              --merch-card-collection-header-filter-padding: 15px;
+              --merch-card-collection-header-sort-height: var(--merch-card-collection-header-filter-height);
+              --merch-card-collection-header-sort-font-size: var(--merch-card-collection-header-filter-font-size);
+              --merch-card-collection-header-sort-padding: var(--merch-card-collection-header-filter-padding);
+              --merch-card-collection-header-result-font-size: 14px;
+          }
+  
+          sp-theme {
+              font-size: inherit;
+          }
+  
+          #header {
+              display: grid;
+              column-gap: var(--merch-card-collection-header-column-gap);
+              row-gap: var(--merch-card-collection-header-row-gap);
+              align-items: center;
+              grid-template-columns: var(--merch-card-collection-header-columns);
+              grid-template-areas: var(--merch-card-collection-header-areas);
+              margin-bottom: var(--merch-card-collection-header-margin-bottom);
+              max-width: var(--merch-card-collection-header-max-width);
+          }
+  
+          #header:empty {
+              margin-bottom: 0;
+          }
+          
+          #search {
+              grid-area: search;
+          }
+  
+          #search sp-search {
+              max-width: var(--merch-card-collection-header-search-max-width);
+              width: 100%;
+          }
+  
+          #filter {
+              grid-area: filter;
+              --mod-actionbutton-edge-to-text: var(--merch-card-collection-header-filter-padding);
+              --mod-actionbutton-height: var(--merch-card-collection-header-filter-height);
+          }
+  
+          #filter slot[name="filtersText"] {
+              font-size: var(--merch-card-collection-header-filter-font-size);
+          }
+  
+          #sort {
+              grid-area: sort;
+              --mod-actionbutton-edge-to-text: var(--merch-card-collection-header-sort-padding);
+              --mod-actionbutton-height: var(--merch-card-collection-header-sort-height);
+          }
+  
+          #sort [slot="label-only"] {
+              font-size: var(--merch-card-collection-header-sort-font-size);
+          }
+  
+          #result {
+              grid-area: result;
+              font-size: var(--merch-card-collection-header-result-font-size);
+          }
+  
+          #result[type="search"][quantity="none"] {
+              font-size: inherit;
+          }
+  
+          #custom {
+              grid-area: custom;
+          }
+  
+          /* tablets */
+          @media screen and ${unsafeCSS(TABLET_UP)} {
+              :host {
+                  --merch-card-collection-header-max-width: auto;
+                  --merch-card-collection-header-columns: 1fr fit-content(100%) fit-content(100%);
+                  --merch-card-collection-header-areas: "search filter sort" 
+                                                        "result result result";
+              }
+          }
+  
+          /* Laptop */
+          @media screen and ${unsafeCSS(DESKTOP_UP)} {
+              :host {
+                  --merch-card-collection-header-columns: 1fr fit-content(100%);
+                  --merch-card-collection-header-areas: "result sort";
+                  --merch-card-collection-header-result-font-size: inherit;
+              }
+          }
+      `;
+  
+      get placeholderKeys() {
+          return [
+              'searchText',
+              'filtersText',
+              'sortText',
+              'popularityText',
+              'alphabeticallyText',
+              'noResultText',
+              'resultText',
+              'resultsText',
+              'resultMobileText',
+              'resultsMobileText',
+              'noSearchResultsText',
+              'searchResultText',
+              'searchResultsText',
+              'noSearchResultsMobileText',
+              'searchResultMobileText',
+              'searchResultsMobileText',
+          ];
+      }
+  }
+  
+  customElements.define('merch-card-collection-header', MerchCardCollectionHeader);  
+
+// #endregion
